feat: keyword search
